refactor(files): extract duplicate functions to shared mod.rs
Move identical functions from fs_private.rs and fs_public.rs to mod.rs
to eliminate code duplication and improve maintainability.

Extracted functions:
- deserialize_data_map: Handles backward compatibility between old (DataMapLevel)
  and new (DataMap) formats for both private and public data maps
- fetch_chunks_parallel: Downloads multiple chunks in parallel using batch
  processing for both private and public downloads

Benefits:
- Reduces code duplication (70+ lines eliminated)
- Single source of truth for data map deserialization logic
- Consistent parallel chunk fetching behavior
- Easier maintenance and bug fixes
- Both private and public downloads now use identical implementation

All tests pass and streaming download functionality remains unchanged.
